// src/gen/clock_pb.ts
// @generated by protoc-gen-es v2.11.0 with parameter "target=ts"
// @generated from file clock.proto (package p2p_sync, syntax proto3)
/* eslint-disable */

import type { GenFile, GenMessage } from "@bufbuild/protobuf/codegenv2";
import { fileDesc, messageDesc } from "@bufbuild/protobuf/codegenv2";
import type { Message } from "@bufbuild/protobuf";

/**
 * Describes the file clock.proto.
 */
export const file_clock: GenFile = /*@__PURE__*/
  fileDesc("CgtjbG9jay5wcm90bxIIcDJwX3N5bmMicgoLVmVjdG9yQ2xvY2sSMwoHZW50cmllcxgBIAMoCzIiLnAycF9zeW5jLlZlY3RvckNsb2NrLkVudHJpZXNFbnRyeRouCgxFbnRyaWVzRW50cnkSCwoDa2V5GAEgASgNEg0KBXZhbHVlGAIgASgNOgI4ASJnCgxOb2RlTWFwRW50cnkSDwoHbm9kZV9uchgBIAEoDRIPCgdub2RlX2lkGAIgASgMEiQKBWNsb2NrGAMgASgLMhUucDJwX3N5bmMuVmVjdG9yQ2xvY2sSDwoHbGVhdmluZxgEIAEoCCJuCgdOb2RlTWFwEicKB2VudHJpZXMYASADKAsyFi5wMnBfc3luYy5Ob2RlTWFwRW50cnkSFAoMbmV4dF9ub2RlX25yGAIgASgNEiQKBWNsb2NrGAMgASgLMhUucDJwX3N5bmMuVmVjdG9yQ2xvY2tiBnByb3RvMw");

/**
 * @generated from message p2p_sync.VectorClock
 */
export type VectorClock = Message<"p2p_sync.VectorClock"> & {
  /**
   * @generated from field: map<uint32, uint32> entries = 1;
   */
  entries: { [key: number]: number };
};

/**
 * Describes the message p2p_sync.VectorClock.
 * Use `create(VectorClockSchema)` to create a new message.
 */
export const VectorClockSchema: GenMessage<VectorClock> = /*@__PURE__*/
  messageDesc(file_clock, 0);

/**
 * @generated from message p2p_sync.NodeMapEntry
 */
export type NodeMapEntry = Message<"p2p_sync.NodeMapEntry"> & {
  /**
   * @generated from field: uint32 node_nr = 1;
   */
  nodeNr: number;

  /**
   * @generated from field: bytes node_id = 2;
   */
  nodeId: Uint8Array;

  /**
   * @generated from field: p2p_sync.VectorClock clock = 3;
   */
  clock?: VectorClock;

  /**
   * @generated from field: bool leaving = 4;
   */
  leaving: boolean;
};

/**
 * Describes the message p2p_sync.NodeMapEntry.
 * Use `create(NodeMapEntrySchema)` to create a new message.
 */
export const NodeMapEntrySchema: GenMessage<NodeMapEntry> = /*@__PURE__*/
  messageDesc(file_clock, 1);

/**
 * @generated from message p2p_sync.NodeMap
 */
export type NodeMap = Message<"p2p_sync.NodeMap"> & {
  /**
   * @generated from field: repeated p2p_sync.NodeMapEntry entries = 1;
   */
  entries: NodeMapEntry[];

  /**
   * @generated from field: uint32 next_node_nr = 2;
   */
  nextNodeNr: number;

  /**
   * @generated from field: p2p_sync.VectorClock clock = 3;
   */
  clock?: VectorClock;
};

/**
 * Describes the message p2p_sync.NodeMap.
 * Use `create(NodeMapSchema)` to create a new message.
 */
export const NodeMapSchema: GenMessage<NodeMap> = /*@__PURE__*/
  messageDesc(file_clock, 2);